fix: all golangci-lint errors; add UPX + separate release archives
- errcheck: handle CloseWrite error in upstream handler
- unconvert: remove uint8(FlagData/Parity) and net.IP() casts
- unparam: readFull returns only error; decodePSK returns only []byte
- gofmt: reformat config.go, chacha.go, ctrl.go

goreleaser: UPX compression (best+lzma), client and server packaged
into separate archives

+++ pkg/transport/downstream/fec.go
@@ -134,7 +134,7 @@ func (e *FECEncoder) flush(partial bool) error {
 
 	// Send data shards.
 	for i := 0; i < actualData; i++ {
-		flags := uint8(FlagData)
+		flags := FlagData
 		if partial {
 			flags |= FlagPartial
 		}
@@ -150,7 +150,7 @@ func (e *FECEncoder) flush(partial bool) error {
 
 	// Send parity shards.
 	for i := 0; i < e.parityShards; i++ {
-		flags := uint8(FlagParity)
+		flags := FlagParity
 		payload := shards[e.dataShards+i]
 		if partial {
 			flags |= FlagPartial
